Devil antag two, electric boogaloo (#2409)
* Baseline


* revival contract & other stuff


* handshake & stuff


* bweh


* posession shitcode
deepseek, save me deepseek :pray:

goob is kil


* other improvements


* devil weakness & stuff


* i forgot to commit the textures...


* power level stuff


* committing so i can push to origin


* bweh


* coderabbit suggestions


* do NOT pass go.
do NOT collect 200.
go STRAIGHT to hell.


* possession actually works, wazow


* genericized le systema

she five on my pebbles until i rainworld


* removed testing ability


* Genericized possession, now you can reuse this for revenants or whatever!


* fixed thingy


* crematorium immune + banishment
remind me to update the guidebook later


* guidebook


* chore: Automatically update REUSE headers

* bweh


* chore: Automatically update REUSE headers

* Update CPRSystem.cs

* polish


* possession & condemned cleanup


* forgot some things :3


* more clauses.
probably works, didnt test :godo:


* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* devil is now actually rollable as a gamemode :fire:


* unshit-coded biblecode


* chore: Automatically update REUSE headers

* mark ur goobcode folks


* chore: Automatically update REUSE headers

* welp.
this is gonna suck


* its all protos now vro


* chore: Automatically update REUSE headers

* minor cleanup


* chore: Automatically update REUSE headers

* even more minor tweaks


* localized that bitch


* remove shitcode


* guh


* fixes asked for by le gus


* new contract color + other fixes


* oopsie


* whoopsie x2


* oopsie x3


* whoopsie x4


* Update Content.Goobstation.Server/Devil/Contract/DevilContractSystem.cs


* chore: Automatically update REUSE headers

* linter STILL bitching


* misc bugs i found


* other bugfixes + part removal is randomized


* other stuff


* reduced hour requirement to five hours of lawyer


* misc fixes


* bweh


* tiny thingy


* teensy bug fix


* little fix


* typo pov


* fuck you


* whoops, forgot about these!


* i put the new forgis on the jeep


* bweh


* wea


* null your entityUids silly


* bweh


* woop


* i want to krill myself


* requested changes


* chore: Automatically update REUSE headers

* fixed shitcode


* chore: Automatically update REUSE headers

* changed requested by ilya (whoops, didnt see these ones!)


* unhardcode that shit


* whoopsie :p


* test fail fix


* chore: Automatically update REUSE headers

* fixed


* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* whoops


* wham


* devil is rollable


* Update Content.Goobstation.Server/Devil/Roles/DevilRoleComponent.cs


* chore: Automatically update REUSE headers

* Update Content.Goobstation.Server/Devil/HandShake/DevilSystem.Handshake.cs


* Update Content.Goobstation.Server/Devil/HandShake/DevilSystem.Handshake.cs


* chore: Automatically update REUSE headers

* requested changes


* issues fixed


* fixed


* this stuff


* most of the changes, but revival contract no worky.


* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* i CANTTT


* im so done


* fixed


* requested changes


* fixed


* webedit duration

* bweh


* whoop


* other stuff


* fix regex


* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* wharg


* Misc bugs


* wowzers, you can actually heal now


* fix test fail


* cheat death tweaks


* wharg


* bad, hardcode no good


* FREE MEE!!!!!


* changed to use shitmeds events instead of brute forcing limb removal


* thaaaanks mocho :))


* Reduced required lawyer time

* Update nukeops.yml

* less stupid i hope


---------


(cherry picked from commit 0fdc118df392e86a58fb34bc3bda20e1dacc0869)

+++ Content.Server/Morgue/CrematoriumSystem.cs
@@ -135,6 +135,9 @@ public sealed class CrematoriumSystem : EntitySystem
             for (var i = storage.Contents.ContainedEntities.Count - 1; i >= 0; i--)
             {
                 var item = storage.Contents.ContainedEntities[i];
+                if (HasComp<CrematoriumImmuneComponent>(item)) // GOOBCODE ALERT //
+                    continue;
+
                 _containers.Remove(item, storage.Contents);
                 EntityManager.DeleteEntity(item);
             }
